pulled erase storage into its own component
