+++ reverb-vscode-webview/src/styles/output.scss
@@ -26,6 +26,7 @@
   .button__terminal {
     width: 4rem;
     height: 1.8rem;
+    border-bottom: 0.05rem solid $border-color;
   }
   .header__metrics {
     display: flex;
